+++ src/main/java/com/omnia/Involutio/entity/WorkerEntity.java
@@ -2,9 +2,7 @@ package com.omnia.Involutio.entity;
 
 import jakarta.persistence.*;
 import lombok.Data;
-import lombok.Generated;
 import lombok.NoArgsConstructor;
-import lombok.extern.java.Log;
 
 
 @Entity
@@ -19,5 +17,5 @@ public class WorkerEntity {
     private String speciality;
     @Column(unique = true)
     private String mail;
-    private Double rating;
+    private Integer rating;
 }
